+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+# Release Notes for 0.1.x
+
+## [v0.2.0 (2023-09-29)](https://github.com/barbosa89/phenix/compare/0.2.0...0.1.0)
+
 ### Added
 - Add `paginate` method to the query builder. ([42](https://github.com/barbosa89/phenix/pull/42))
 - Add `count` method to the query builder. ([42](https://github.com/barbosa89/phenix/pull/42))
 - Add `insert` method to the query builder. ([43](https://github.com/barbosa89/phenix/pull/43))
 - Add `exists` and `doesntExists` methods to the query builder. ([#44](https://github.com/barbosa89/phenix/pull/44))
+- Add `delete` method to the query builder. ([#45](https://github.com/barbosa89/phenix/pull/45))
 
 ### Changed
 - Load routes before server running. ([#41](https://github.com/barbosa89/phenix/pull/41))
@@ -19,7 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve service provider structure. ([#38](https://github.com/barbosa89/phenix/pull/38))
 - Improve class API to `\Core\Database\QueryGenerator`, now it has final methods. ([#44](https://github.com/barbosa89/phenix/pull/44))
 
-# Release Notes for 0.1.x
+### Fixed
+- Apply provides in database service provider. ([#46](https://github.com/barbosa89/phenix/pull/46))
 
 ## [v0.1.0 (2023-09-15)](https://github.com/barbosa89/phenix/compare/0.1.0...0.0.1-alpha.1)